more efficient expression generator
--HG--
extra : convert_revision : svn%3A04d2d0ec-5548-0410-87ee-dd41c663a700/trunk%40601

+++ hosc/src/hosc/TypeAlgebra.scala
@@ -3,7 +3,7 @@ package hosc
 import EnrichedLambdaCalculus._
 
 object TypeAlgebra {
-  private var n: Int = 0
+  private var n: Long = 0
   def newTyvar(): TypeVariable = { n += 1; TypeVariable("$$" + n) }
   
   def tyvars(t: Type): List[TypeVariable] = t match {
